better loading notification on mission events page
